+ Clock in MenuScene but BUG BUG BUG

+++ trunk/Classes/Global.h
@@ -45,7 +45,7 @@ extern float G_ITEM_BULLET_RANDOM_PERCENT;
 extern float G_ITEM_ARMOR_RANDOM_PERCENT;
 extern float G_ITEM_BOOM_RANDOM_PERCENT;
 
-extern float G_PLAYER_TIME_TO_REVIVE;
+extern int G_PLAYER_TIME_TO_REVIVE;
 extern int G_MAX_PLAYER_LIFE; 
 
 extern int G_MAX_PLAYER_BOOM;
